train darcy with refactored comm
